+++ src/liberator/dev.clj
@@ -2,7 +2,6 @@
   (:use hiccup.core
         hiccup.page
         compojure.core
-        clojure.tools.trace
         [liberator.core :only [defresource]]
         [clojure.string :only [join]])
   (:require [liberator.core :as core]
@@ -232,7 +231,7 @@
                (do
                  (save-log! *current-id*
                             [(Date.)
-                             (select-keys request [:request-method :uri :headers])
+                             (select-keys request [:request-method :uri :headers :params])
                              @request-log])
                  (assoc-in resp [:headers trace-id-header] *current-id*))
                resp))))))