+++ src/liberator/representation.clj
@@ -209,6 +209,10 @@
   (as-response [this context]
     (as-response (render-map-generic this context) context))
 
+  clojure.lang.MultiFn
+  (as-response [multi-fn context]
+    (as-response (multi-fn context) context))
+
   ;; If a string is returned, we should carry out the conversion of both the charset and the encoding.
   String
   (as-response [this {representation :representation}]
@@ -238,4 +242,3 @@
     response))
 
 (defn ring-response [map] (->RingResponse map))
-